fix: extracted pygame out of main.py
